change configuration of some functions in repositories and combined some content repositories
