--- app/src/main/java/com/example/taskplanner/domain/repository/auth/signin/SignInRepository.kt
@@ -1,8 +1,0 @@
-package com.example.taskplanner.domain.repository.auth.signin
-
-import com.example.taskplanner.util.Resources
-import com.google.firebase.auth.AuthResult
-
-interface SignInRepository {
-    suspend fun signIn(email: String, password: String): Resources<AuthResult>
-}
