--- app/src/main/java/com/example/taskplanner/domain/repository/auth/signup/SignUpRepository.kt
@@ -1,9 +1,0 @@
-package com.example.taskplanner.domain.repository.auth.signup
-
-import com.example.taskplanner.domain.model.UserDomain
-import com.example.taskplanner.util.Resources
-import com.google.firebase.auth.AuthResult
-
-interface SignUpRepository {
-    suspend fun signUp(userDomain: UserDomain):Resources<AuthResult>
-}
