--- app/src/main/java/com/example/taskplanner/domain/repository/project/create/ProjectCreatorRepository.kt
@@ -1,8 +1,0 @@
-package com.example.taskplanner.domain.repository.project.create
-
-import com.example.taskplanner.domain.model.ProjectDomain
-import com.example.taskplanner.util.Resources
-
-interface ProjectCreatorRepository {
-    suspend fun createProject(projectDomain: ProjectDomain): Resources<Unit>
-}
