--- app/src/main/java/com/example/taskplanner/domain/repository/task/details/TaskDetailsRepository.kt
@@ -1,12 +1,0 @@
-package com.example.taskplanner.domain.repository.task.details
-
-import com.example.taskplanner.domain.model.TaskDomain
-import com.example.taskplanner.util.Progress
-import com.example.taskplanner.util.Resources
-
-interface TaskDetailsRepository {
-    suspend fun getTaskInfo(taskId: String): Resources<TaskDomain>
-    suspend fun updateTask(projectId: String, taskDomain: TaskDomain, ): Resources<Unit>
-    suspend fun updateTaskProgress(taskId: String, progress: Progress): Resources<Unit>
-    suspend fun deleteTask(taskId: String): Resources<Unit>
-}
