--- app/src/main/java/com/example/taskplanner/util/Constants.kt
@@ -1,14 +1,0 @@
-package com.example.taskplanner.util
-
-object Constants {
-    const val USER_COLLECTION = "user"
-    const val PROJECT_COLLECTION = "project"
-    const val TASK_COLLECTION = "task"
-    const val PROJECTS_OWNER_ID = "ownerId"
-    const val TITLE_FIELD = "title"
-    const val DESCRIPTION_FIELD = "description"
-    const val START_TIME_FIELD = "startTime"
-    const val END_TIME_FIELD = "endTime"
-    const val TASK_PROGRESS_FIELD = "taskProgress"
-    const val PROJECT_PROGRESS_FIELD = "projectProgress"
-}
